+++ virtual-pipeline/src/main/java/io/jenkins/plugins/VirtualPipeline/VirtualPipelineOutputHistoryMarked.java
@@ -7,7 +7,7 @@ public class VirtualPipelineOutputHistoryMarked extends VirtualPipelineLineOutpu
 
     private HistoryType historyType;
 
-    public VirtualPipelineOutputHistoryMarked(String regex, String line, int index, Boolean deleteMark, LineType type, HistoryType historyType, int lineOffset) {
+    public VirtualPipelineOutputHistoryMarked(String regex, String line, int index, Boolean deleteMark, LineType type, HistoryType historyType, long lineOffset) {
         super(regex, line, index, deleteMark, type, lineOffset);
         this.historyType = historyType;
     }
